requirements.txt
flask
gunicorn
slackclient
python-dateutil
slackeventsapi
pypco
